Struct solana_thin_client::thin_client::ThinClient

pub struct ThinClient<P, M, C> {
    rpc_clients: Vec<RpcClient>,
    tpu_addrs: Vec<SocketAddr>,
    optimizer: ClientOptimizer,
    connection_cache: Arc<ConnectionCache<P, M, C>>,
}

An object for querying and sending transactions to the network.

Fields
rpc_clients: Vec<RpcClient>tpu_addrs: Vec<SocketAddr>optimizer: ClientOptimizerconnection_cache: Arc<ConnectionCache<P, M, C>>
Implementations

impl<P, M, C> ThinClient<P, M, C>where P: ConnectionPool<NewConnectionConfig = C>, M: ConnectionManager<ConnectionPool = P, NewConnectionConfig = C>, C: NewConnectionConfig,

pub fn new( rpc_addr: SocketAddr, tpu_addr: SocketAddr, connection_cache: Arc<ConnectionCache<P, M, C>> ) -> Self

Create a new ThinClient that will interface with the Rpc at rpc_addr using TCP and the Tpu at tpu_addr over transactions_socket using Quic or UDP (currently hardcoded to UDP)

pub fn new_socket_with_timeout( rpc_addr: SocketAddr, tpu_addr: SocketAddr, timeout: Duration, connection_cache: Arc<ConnectionCache<P, M, C>> ) -> Self

fn new_from_client( rpc_client: RpcClient, tpu_addr: SocketAddr, connection_cache: Arc<ConnectionCache<P, M, C>> ) -> Self

pub fn new_from_addrs( rpc_addrs: Vec<SocketAddr>, tpu_addrs: Vec<SocketAddr>, connection_cache: Arc<ConnectionCache<P, M, C>> ) -> Self

fn tpu_addr(&self) -> &SocketAddr

pub fn rpc_client(&self) -> &RpcClient

pub fn retry_transfer_until_confirmed( &self, keypair: &Keypair, transaction: &mut Transaction, tries: usize, min_confirmed_blocks: usize ) -> TransportResult<Signature>

Retry a sending a signed Transaction to the server for processing.

pub fn retry_transfer( &self, keypair: &Keypair, transaction: &mut Transaction, tries: usize ) -> TransportResult<Signature>

Retry sending a signed Transaction with one signing Keypair to the server for processing.

pub fn send_and_confirm_transaction<T: Signers + ?Sized>( &self, keypairs: &T, transaction: &mut Transaction, tries: usize, pending_confirmations: usize ) -> TransportResult<Signature>

pub fn poll_get_balance(&self, pubkey: &Pubkey) -> TransportResult<u64>

pub fn poll_get_balance_with_commitment( &self, pubkey: &Pubkey, commitment_config: CommitmentConfig ) -> TransportResult<u64>

pub fn wait_for_balance( &self, pubkey: &Pubkey, expected_balance: Option<u64> ) -> Option<u64>

pub fn get_program_accounts_with_config( &self, pubkey: &Pubkey, config: RpcProgramAccountsConfig ) -> TransportResult<Vec<(Pubkey, Account)>>

pub fn wait_for_balance_with_commitment( &self, pubkey: &Pubkey, expected_balance: Option<u64>, commitment_config: CommitmentConfig ) -> Option<u64>

pub fn poll_for_signature_with_commitment( &self, signature: &Signature, commitment_config: CommitmentConfig ) -> TransportResult<()>

pub fn get_num_blocks_since_signature_confirmation( &mut self, sig: &Signature ) -> TransportResult<usize>